+++ tests/utils/transaction.js
@@ -15,10 +15,9 @@ exports.generateExampleTransactionsJson = function (n = 1) {
     let transactions = [];
     let exampleTransaction = JSON.parse(JSON.stringify(defaultExampleTransaction));
     for (let i = 0; i < n; i++) {
-        let transactionAmount = random.randomInt(1, 10);
         exampleTransaction.receiverId = '5768064935133184';
-        exampleTransaction.amount = transactionAmount;
-        exampleTransaction.externalId = 'Test ' + Date();
+        exampleTransaction.amount = random.randomInt(1, 10);
+        exampleTransaction.externalId = random.randomInt(1, 1.e16).toString();
         transactions.push(Object.assign(new starkbank.Transfer({}), JSON.parse(JSON.stringify(exampleTransaction))));
     }
     return transactions;