+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Change
+- Test user credentials to environment variable instead of hard-code
 
 ## [0.4.0] - 2020-05-12
 ### Added